pyproject.toml
[tool.poetry]
name = "devkami-generator"
version = "0.1.0"
description = ""
authors = ["sweemeng <sweester@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
google-api-python-client = "^1.12.1"
google-auth-oauthlib = "^0.4.1"
google-auth-httplib2 = "^0.0.4"
dateutils = "^0.6.11"
python-slugify = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
